+++ HISTORY.md
@@ -2,5 +2,5 @@
 
 ## Version 0.1.0
 
-- Release date: TBD
-- Db Connect module: gorm
+- Release date: Dec 05, 2020
+- Db Connect module: for RDBMS => PostgresDB, MySQL, SQLite